refactor(storage): extract _JobsMixin
Moves 7 rescan jobs queue methods (enqueue_rescan_job, claim_next_queued_job,
requeue_stale_running_jobs, mark_job_done, mark_job_failed,
get_rescan_queue_summary, get_job) to storage/jobs.py with its own module
logger (mark_job_done/mark_job_failed emit warnings on lost claims).
